add intermediate save_model

+++ finetune/datasets.py
@@ -42,8 +42,8 @@ class COCOFlickrDataset(Dataset):
         image = Image.open(self.get_img_path(idx))
         caption = self.annotations[idx]["caption"]
         # image = self.image_processor([[image]]).squeeze(0)
-        image = self.image_processor([[image]]).half().squeeze(0)
-
+        # image = self.image_processor([[image]]).half().squeeze(0)
+        image = self.image_processor([[image]]).squeeze(0).to(torch.bfloat16)
         attention_mask, input_ids, labels = self._process_text(caption)
         attention_mask, input_ids, labels = self._pad_text(attention_mask, input_ids, labels)
         return image, input_ids, labels, attention_mask
